feat(output/lineno): show the line number if fail
